add stick to top btn
