+++ package.json
@@ -7,6 +7,7 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^14.4.3",
+    "aos": "^2.3.4",
     "axios": "^1.2.3",
     "daisyui": "^2.47.0",
     "node-fetch": "^3.3.0",
