+++ src/App.css
@@ -1,0 +1,18 @@
+.scroll_up {
+    right: 17px;
+    bottom: 20px;
+    width: 45px;
+    height: 45px;
+    z-index: 99;
+    align-items: center;
+    justify-content: center;
+    border-radius: 50%;
+    transition: all 0.3s ease-in-out;
+    box-shadow: 0 0 25px rgba(0, 0, 0, 0.1);
+    font-size: 25px;
+    font-weight: 500;
+}
+
+ .scroll_up:hover {
+    transform: translateY(-4px);
+}
