--- src/App.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import './App.css';
-import Routes from "./router/Routes"
-
-function App() {
-  return (
-    <>
-    <Routes />
-    </>
-  );
-}
-
-export default App;
