+++ src/pages/HomePage.jsx
@@ -1,9 +1,9 @@
-import Brands from "../components/Brands";
-import Cta from "../components/Cta";
-import MainServices from "../components/MainServices";
-import MainCategories from "../components/MainCategories";
-import MainMedServices from "../components/MainMedServices";
-import MainProducts from "../components/MainProducts";
+import Brands from "../components/home/Brands";
+import Cta from "../components/home/Cta";
+import MainServices from "../components/home/MainServices";
+import MainCategories from "../components/home/MainCategories";
+import MainMedServices from "../components/home/MainMedServices";
+import MainProducts from "../components/home/MainProducts";
 import Slider from "../components/Slider";
 import MainLayout from "../Layout/MainLayout";
